// src/app/models/letter/ORMLetterSections.ts
export class ORMLetterSections {
    section_id: String;
    section_name: String;
    section_text: String;
    practice_id: String;
    deleted: Boolean;
    created_user: String;
    client_date_created: String;
    modified_user: String;
    client_date_modified: String;
    date_created: String;
    date_modified: String;
    system_ip: String;
}